Added cluster move; Cylinder with periodic ends; Fixed problem NPT gsize where the average counter would turn for long runs.

+++ include/faunus/average.h
@@ -47,7 +47,7 @@ namespace Faunus {
   
   template<class T> Average<T>::Average() { reset(); }
   
-  template<class T> T Average<T>::avg() const { return sum/cnt; }
+  template<class T> T Average<T>::avg() const { return double(sum)/cnt; }
 
   template<class T> void Average<T>::reset() {
     sum=sqsum=0;
